requirements.txt
yt-dlp>=2024.8.6
python-dotenv>=1.0.1
customtkinter>=5.2.2
pytube>=15.0.0
pyperclip>=1.8.2
pystray>=0.19.5
plyer>=2.1.0
Pillow>=10.0.0
